Update marketplace-clean.ps1 so it can keep git files

+++ run-command/marketplace-clean.ps1
@@ -1,6 +1,14 @@
+param
+(
+    [bool]$KeepGit = $false
+)
 
 Get-childitem -recurse -filter ".vs*" -Force | Remove-Item -Force -Recurse -Verbose
-Get-childitem -recurse -filter ".git*" -Force | Remove-Item -Force -Recurse -Verbose
-Remove-Item -Recurse -Force ./Intermediate
-Remove-Item -Recurse -Force ./Binaries  
+if(!$KeepGit)
+{
+  Get-childitem -recurse -filter ".git*" -Force | Remove-Item -Force -Recurse -Verbose
+}
+
+Remove-Item -Recurse -Force ./Intermediate -ErrorAction Ignore
+Remove-Item -Recurse -Force ./Binaries -ErrorAction Ignore
 Remove-Item -Recurse -Force ./*metadata -ErrorAction Ignore
